Restructure tasks dir: one subdirectory per task with task.py
Each of the 7 ML task modules is moved from a flat .py file into its own
subdirectory of the same name, with the module file renamed to task.py and
an empty __init__.py alongside it. This gives each task a natural home for
future helpers, schemas, or fixtures without cluttering the shared directory.

Updated files:
- celery.py and tasks_validator.py: module path template changed from
  'starfish.controller.tasks.{name}' to 'starfish.controller.tasks.{name}.task'
- All 7 test files: import paths updated to match new structure

Closes #18
