+++ controller/starfish/controller/test_logistic_regression_stats.py
@@ -10,7 +10,7 @@ from unittest.mock import patch
 import statsmodels.api as sm
 from sklearn.model_selection import train_test_split
 
-from starfish.controller.tasks.logistic_regression_stats import (
+from starfish.controller.tasks.logistic_regression_stats.task import (
     LogisticRegressionStats,
 )
